refactor: Project, Task 타입 선언 분리 및 import 최신화 [#100]
