+++ client/src/types/project.d.ts
@@ -1,0 +1,4 @@
+export type ProjectType = {
+  id: number;
+  name: string;
+};
